Split nft container to components
