+++ components/nft/index.ts
@@ -1,2 +1,4 @@
 export { default as NftCard } from './NftCard'
 export { default as NftDetails } from './NftDetails'
+export { default as NftPreview } from './NftPreview'
+export { default as NftMinted } from './NftMinted'
